ldcm: Longitudinal Diagnostic Classification Models — ldcm-package

Diagnostic classification models (DCMs) can provide useful information about students' strengths and weaknesses in an academic domain beyond the norm-referenced or percentile results typically achieved through item response theory. DCMs, however, need to be able to accommodate the multilevel framework in which most education takes place, or they run the risk of providing substantially inaccurate results. The purpose of this project is to develop and make accessible to applied researchers a multilevel extension to the longitudinal diagnostic classification model (DCM) to help researchers take into account contextual effects that can impact the fidelity and effectiveness of an educational intervention.

Details

See vignette("ldcm", package = "ldcm") for an overview of the package.

See also

Author

Maintainer: Michael E. Cotterell mepcott@uga.edu (ORCID) [copyright holder]

Authors:

Matthew Madison mjmadison@uga.edu [copyright holder]

Minjeong Jeon mjjeon@ucla.edu [copyright holder]

Other contributors:

University of Georgia (The content and opinions expressed in this material do not necessarily reflect the views of nor are they endorsed by the University of Georgia or the University System of Georgia.) [copyright holder]

U.S. Department of Education Institute of Education Sciences (This work is supported by the U.S. Department of Education Institute of Education Sciences under IES Award Number R305D220020. Any opinions, findings, conclusions or recommendations expressed in this material are those of the author(s) and do not necessarily reflect those of the Institute of Educational Sciences or the U.S. Department of Education.) [funder]